fix bugs with attaching holes to containing faces.
when connecting two or more holes to the same face vertex, it was
possible to cause a twist in the face loop, which subsequently
messed up triangulation. the trick to fixing this is to make sure
that the hole vertex to be joined to the face vertex is within the
cone made by the vertex it is to be joined to, and the prior and
next face vertices. this necessitated a refactoring of the
candidate attachment testing code.

N.B. this same bug exists in the templated version of the code in
include/carve/triangulation_impl.hpp. it needs to be fixed there
too.