fix the division of faces by the line of intersection.
this change modifies the way that the line of intersection is merged
into a split face loop. originally, this was done in all cases by
tracing paths in the graph of face edges and inserted intersection
edges. because determining the direction to head at a branch point
was based upon angle comparisons, this could be confused by colinear
edges.

now the procedure is as follows:

collect intersection edges into paths and loops by forming a graph
from the edges, and tracing connected components that end at a
branch point, or touching the face loop. loops never branch, and
do not intersect the face loop, so they immediately form holes.
paths that begin and end on the face loop divide the face in two,
so they can be dealt with immediately (this fixes the colinear edge
problem [in all but particularly morbid cases, i think]).

in the case where there is a mixture of edge types, the crossing
edges are used to divide up the face as far as possible, and then
non-crossing edges and loops are assigned to one of the divided
face loops, and the original graph tracing code is executed to
process them.

in most cases, however, where a single edge cuts a face in two,
almost all of this can be avoided (previously the graph traversal
code was still called, even in this trivial case).

care must now be taken when an edge that doubles back on itself
(e.g.  where a hole has previously been joined to a face) is
intersected.  in this case the intersection vertex is repeated in
the face loop, and the index of the face loop that the hole attaches
to must be correctly determined using an internal-to-angle test,
as per hole connection. previously this wasn't a problem, as the
graph traversal code forgets about the ordering of vertices in the
face loop, and instead decomposes it into individual edges.