Correct edge classification in the case of multiple touching manifolds.
For example, take two cubes that share a face. A face intersecting
this pair of manifolds will be classified as inside one, but outside
the other. The correct choice is to classify the face as inside.
Doing this means that cases where manifold embedding is inconsistent
will not be detected as such, but maybe this problem should be detected
and handled elsewhere, anyway.