Respect classification against both closed and open meshes.
If classification wrt. a closed mesh is available, choose that.
Otherwise, accept classification against an open mesh. The edge
classififer uses local topology at the line of intersection to
determine orientation, rather than relying on point-in-polyhedron
tests, and so the information it returns for intersections with
open meshes is still useful.