Provide a MeshOptions struct for passing mesh creation options.
Currently there is only one option - avoid_cavities, which determines
which direction to take when tracing faces in order to resolve
complex edges. If avoid_cavities is true, cavities that share an
edge with an enclosing mesh will be joined to that mesh. If it is
false (the default) then cavities are represented as separate meshes.